added a plot base class
